jj-fzf/jj-fzf: fixup `jj squash {-r|--from} @`
The behavior for `jj squash` of @ is not desirable at all.
What we need is for the squashed commit to retain change_id, children,
merge-parents, bookmarks, but move the message along with the diff.
Using --keep-emptied does not move the message, but without it,
all of the above is lost, so for now resort to doing
`jj new --insert-before @` before the squash (which will receive
any children, merge-parents and bookmarks) and never ever
squash @ directly.

Signed-off-by: Tim Janik <timj@gnu.org>